# Optimize string argsort