Skip checking if strings are already sorted and use more aggregation to
optimize string argsort

`SegString.isSorted()` involves a lot of fine-grained remote array
accesses, which doesn't perform well. Skip that check since it's faster
to just sort directly. We can probably optimize this check later, but
for now it was easier to just skip it.

Use aggregation to gather values in string sorting. Elements from the
`values` array can be remote, so using aggregation avoids lots of
fine-grained remote accesses. To do that replace `unorderedCopy` for the
first `copyDigit` with assignment for local operations and aggregation
for the potentially remote `values`. Increase aggregation usage for the
second `copyDigit` by aggregating the remote `values` as well. Note that
this requires calculating and storing the bucket position prior to
gathering the new values/digits.

Perf results for str-argsort on 16-node-xc and 16-node-cs-hdr. First 3
use the current default problem size (10**6), but with this we can start
running larger problem sizes as shown in the last 2 rows:

| config            | 16-node-xc   | 16-node-cs-hdr |
| ----------------- | ------------ | -------------- |
| master            | 0.0086 GiB/s | 0.0027 GiB/s   |
| no checkSorted    | 0.0414 GiB/s | 0.0035 GiB/s   |
| copyDigit agg     | 0.1137 GiB/s | 0.0118 GiB/s   |
| --size=$((10**8)) | 0.8214 GiB/s | 0.7413 GiB/s   |
| --size=$((2**29)) | 0.8543 GiB/s | 1.6285 GiB/s   |

Part of https://github.com/mhmerrill/arkouda/issues/659